% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.anna-package.R
\docType{package}
\name{data.anna-package}
\alias{data.anna}
\alias{data.anna-package}
\title{data.anna: Build A Package With Internal And External Data}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org}

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
