% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_six_from_col.R
\name{extract_six_from_col}
\alias{extract_six_from_col}
\title{extract_six_from_col A function to sort decreasing and extract the first 6 elements of a given column}
\usage{
extract_six_from_col(dataset = datasets::mtcars, column = "mpg")
}
\arguments{
\item{dataset}{dataset given to extract a column from (type = tibble, data.frame)}

\item{column}{column to sort and extract the first 6 elements from (type = numeric, character)}
}
\value{
A data frame when the column is numeric
}
\description{
extract_six_from_col A function to sort decreasing and extract the first 6 elements of a given column
}
\examples{
extract_six_from_col(dataset = datasets::mtcars, column = "mpg")
extract_six_from_col(squirrels, "lat") 
}
